scripts/requirements.txt
GitPython==3.1.30
python-dateutil==2.7.3
PyYAML==5.3.1
requests==2.22.0